+++ app/src/main/java/ua/com/radiokot/photoprism/features/viewer/di/MediaViewerFeatureModule.kt
@@ -20,7 +20,6 @@ import ua.com.radiokot.photoprism.di.INTERNAL_DOWNLOADS_DIRECTORY
 import ua.com.radiokot.photoprism.di.VIDEO_CACHE_DIRECTORY
 import ua.com.radiokot.photoprism.env.data.model.EnvSession
 import ua.com.radiokot.photoprism.features.gallery.di.galleryFeatureModule
-import ua.com.radiokot.photoprism.features.viewer.logic.ArchiveGalleryMediaUseCase
 import ua.com.radiokot.photoprism.features.viewer.logic.BackgroundMediaFileDownloadManager
 import ua.com.radiokot.photoprism.features.viewer.logic.DefaultVideoPlayerFactory
 import ua.com.radiokot.photoprism.features.viewer.logic.DeleteGalleryMediaUseCase
@@ -94,7 +93,6 @@ val mediaViewerFeatureModule = module {
         viewModelOf(::VideoPlayerCacheViewModel)
 
         scopedOf(SetGalleryMediaFavoriteUseCase::Factory)
-        scopedOf(ArchiveGalleryMediaUseCase::Factory)
         scopedOf(DeleteGalleryMediaUseCase::Factory)
     }
 }