+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Ability to archive or delete a single item from the media viewer
+- Ability to archive or delete a single gallery item from the media viewer
+- Ability to archive gallery items selected with a long press
 
 ## [1.26.0] - 2024-04-21